requirements.txt
# Core Flask Dependencies
flask==2.3.3
flask-cors==4.0.0

# Essential Data Processing
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
catboost==1.2.8

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Basic ML Libraries
scipy==1.11.1
matplotlib==3.7.2

# File Processing
openpyxl==3.1.2

# Date and Time
python-dateutil==2.8.2

# Configuration
pyyaml==6.0.1

# Production Server
gunicorn==21.2.0

# Additional Dependencies for Production
Werkzeug==2.3.8
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
